+++ components/logiVehicle/zen_modules.sqf
@@ -1,7 +1,10 @@
+call
+{
+    [
+        "[CA2] Logistics",
+        "Create / restock logi vehicle",
+        f_fnc_zen_makeLogiVehicle
 
-[
-    "[CA2] Logistics",
-    "Create / restock logi vehicle",
-    f_fnc_zen_makeLogiVehicle
-
-] call zen_custom_modules_fnc_register;
+    ] call zen_custom_modules_fnc_register;
+    
+};
