+++ components/miscShared/fn_removeJammersFromObject.sqf
@@ -1,0 +1,7 @@
+/*
+	Created By: Gibbs
+*/
+
+params ["_object"];
+
+[_object, 0] call kyk_ew_fnc_jammerArrayRemove;
