+++ components/miscShared/functions.hpp
@@ -23,7 +23,6 @@ class miscShared
     class factionToSideName{};
     class reapplyInsigniaLocal{};
     class getObjectUid{};
-    class createPlayerSpawner{};
     class addObjectsToAllZeuses{};
     class addCustomSettings{};
     class canPlayerOfSideRespawn{};
@@ -33,4 +32,6 @@ class miscShared
     class getAllObjectVariablesInNamespace{};
     class getUnitConfiguredSide{};
     class woundUnitRandomly{};
+    class addJammerToObject{};
+    class removeJammersFromObject{};
 };
