+++ components/respawnWaves/fn_createPlayerSpawner.sqf
