+++ components/respawnWaves/functions.hpp
@@ -5,9 +5,18 @@ class respawnWaves
     class forceRespawnWave{};
     class respawnWaveServer{};
     class respawnDialogSupport{};
+    class createPlayerSpawner{};
 };
 class respawnWaves_ui
 {
     file = "components\respawnWaves\ui_functions";
     class respawnTitle_onLoad{};
+};
+class respawnWaves_zen
+{
+    file = "components\respawnWaves\zen";
+	class zen_createSpawner{};
+	class zen_moveRespawnMarker{};
+	class zen_setRespawnObject{};
+	class zen_createPlayerSpawner{};
 };
