+++ components/respawnWaves/zen/fn_zen_createPlayerSpawner.sqf
@@ -1,5 +1,7 @@
 #include "macros.hpp"
 
+SERVER_ONLY;
+
 params ["_position", ["_side", west]];
 
 _spawner = [_position, _side] call f_fnc_createPlayerSpawner;
