+++ components/respawnWaves/zen/fn_zen_moveRespawnMarker.sqf
@@ -1,5 +1,7 @@
 #include "macros.hpp"
 
+CLIENT_ONLY;
+
 params ["_position", ["_side", west]];
 
 _sideStr = str _side;
@@ -10,10 +12,10 @@ if (getMarkerColor _markerName isEqualTo "") then
     _marker = createMarker [_markerName, _position];
     _marker setMarkerAlpha 0;
 
-    systemChat format ["Created marker '%1' at given position.", _markerName];
+    ["Created marker '%1' at given position.", _markerName] call zen_common_fnc_showMessage;
 }
 else
 {
     _markerName setMarkerPos _position;
-    systemChat format ["Moved marker '%1' to given position.", _markerName];
+    ["Moved marker '%1' to given position.", _markerName] call zen_common_fnc_showMessage;
 };
