+++ components/respawnWaves/zen/fn_zen_setRespawnObject.sqf
@@ -1,10 +1,12 @@
 #include "macros.hpp"
 
+CLIENT_ONLY;
+
 params ["_position", "_object", ["_side", west]];
 
 if (isNull _object) then
 {
-    systemChat "No object found: respawner object unchanged.";
+    ["No object found: respawner object unchanged."] call zen_common_fnc_showMessage;
 }
 else
 {
@@ -12,6 +14,6 @@ else
     _objectName = toLower ("respawner_" + _sideStr);
 
     missionNamespace setVariable [_objectName, _object, true];
-    systemChat format ["Set %1 respawner object to '%2'.", _sideStr, typeOf _object];
-    
+    ["Set %1 respawner object to '%2'.", _sideStr, typeOf _object] call zen_common_fnc_showMessage;
+
 };
