--- components/zenFunctions/client/macros.hpp
@@ -1,1 +1,0 @@
-#include "../macros.hpp"
