--- components/zenFunctions/functions.hpp
@@ -1,17 +1,0 @@
-class zenFunctions_client
-{
-	file = "components\zenFunctions\client";
-	class zen_createSpawner{};
-	class zen_moveRespawnMarker{};
-	class zen_setRespawnObject{};
-	class zen_removeJammers{};
-	class zen_createJammer{};
-	class zen_toggleLambsReinforcementGroup{};
-};
-class zenFunctions_server
-{
-	file = "components\zenFunctions\server";
-	class zen_createPlayerSpawner{};
-	class zen_doRemoveJammers{};
-	class zen_doCreateJammer{};
-};
