--- components/zenFunctions/server/fn_zen_doRemoveJammers.sqf
@@ -1,8 +1,0 @@
-/*
-	Created By: Gibbs
-*/
-#include "macros.hpp"
-
-params ["_object"];
-
-[_object, 0] call kyk_ew_fnc_jammerArrayRemove;
