--- components/zenFunctions/server/macros.hpp
@@ -1,1 +1,0 @@
-#include "../macros.hpp"
