+++ components/zen_modules.sqf
@@ -1,2 +1,3 @@
 #include "logiVehicle\zen_modules.sqf"
 #include "miscClient\zen_modules.sqf"
+#include "respawnWaves\zen_modules.sqf"
